% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_system_default_font}
\alias{get_system_default_font}
\title{Internal function: Get system default Chinese font (fallback solution)}
\usage{
get_system_default_font()
}
\description{
Internal function: Get system default Chinese font (fallback solution)
}
