% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.scale_fill_color_ctc_d.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First argument.}

\item{b}{Second argument (default value if \code{a} is NULL).}
}
\value{
The value of \code{a} if it is not NULL, otherwise the value of \code{b}.
}
\description{
Returns the first argument if it is not NULL, otherwise returns the second argument.
}
\keyword{internal}
