% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.scale_fill_color_ctc_c.R
\name{scale_color_ctc_c}
\alias{scale_color_ctc_c}
\title{Alias for scale_colour_ctc_c (American English Spelling)}
\usage{
scale_color_ctc_c(palette_name, direction = 1, ...)
}
\arguments{
\item{palette_name}{For built-in palette, the input as below are available:
\enumerate{
\item the name of palette(the element name for a palette in the palette_list),such as "seq01","div18",...;
\item index of the palette, 01-60;
\item the palette_name value in a palette, such as "紫韵花影",for Chinese name;
\item the palette_name_e value in a palette, such as "violet_bloom",for English name;
You can get the palette_name values and a quick preview the palettes by run \code{list_palettes()} function.
}}

\item{direction}{Numeric (1 or -1) to control color order: 1 for original order, -1 for reversed. Defaults to 1.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_color_gradientn()} (e.g., \code{name} for legend title).}
}
\description{
Convenience alias for \code{scale_colour_ctc_c()} to support American English.
}
