% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.scale_fill_color_ctc_c.R
\name{scale_fill_ctc_c}
\alias{scale_fill_ctc_c}
\title{Continuous Fill Color Scale for ggplot2 Using Built-in CTC Palettes}
\usage{
scale_fill_ctc_c(palette_name, direction = 1, ...)
}
\arguments{
\item{palette_name}{For built-in palette, the input as below are available:
\enumerate{
\item the name of palette(the element name for a palette in the palette_list),such as "seq01","div18",...;
\item index of the palette, 01-60;
\item the palette_name value in a palette, such as "紫韵花影",for Chinese name;
\item the palette_name_e value in a palette, such as "violet_bloom",for English name;
You can get the palette_name values and a quick preview the palettes by run \code{list_palettes()} function.
}}

\item{direction}{Numeric (1 or -1) to control color order: 1 for original order, -1 for reversed. Defaults to 1.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_gradientn()} (e.g., \code{name} for legend title).}
}
\value{
A ggplot2 continuous fill color scale object.
}
\description{
Creates a continuous outline color scale for ggplot2 by retrieving colors from built-in CTC palettes
(seq01-seq20, div01-div20, qual01-qual20) via \code{ctc_palette()}. Designed for mapping continuous data
with smooth gradient interpolation.
}
\examples{
\dontrun{
data(palette_list)  # Load built-in palettes

# 1 Use diverging palette for fill,
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
  geom_raster() +
  scale_fill_ctc_c(palette_name = "div12", direction = -1, name = "Density")

  # 2 palette by index value
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
geom_raster() +
    scale_fill_ctc_c(palette_name = 33, direction = 1, name = "Density")

  # 3 palette by Chinese palette_name value
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
    geom_raster() +
    scale_fill_ctc_c(palette_name = "海天沙影", direction = 1, name = "Density")

  # 4 palette by English palette_name_e value
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
    geom_raster() +
    scale_fill_ctc_c(palette_name = "autumn_blue", direction = -1, name = "Density")
}

}
\seealso{
\code{ctc_palette()} for palette generation;
\code{scale_colour_ctc_c()} for fill scales;
\code{plot_palettes()} to visualize valid palettes.
\code{list_palettes()} to display a searchable table of all available color palettes with their identifiers with color preview.
}
